+++ .gitignore
@@ -1,3 +1,6 @@
 node_modules
 .vscode
 .DS_Store
+dist
+docs
+coverage
