+++ __test__/test.test.js
@@ -1,0 +1,1 @@
+test ('the tester is working', () => {expect(true).toBe(true);})
