+++ package.json
@@ -4,9 +4,26 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --coverage",
     "start": "webpack serve --open",
-    "build": "webpack"
+    "build": "webpack",
+    "webhint": "hint ."
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "jsx"
+    ],
+    
+    "moduleDirectories": [
+      "node_modules",
+      "bower_components",
+      "shared"
+    ],
+    "moduleNameMapper": {
+      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js",
+      "\\.(gif|ttf|eot|svg)$": "<rootDir>/__mocks__/fileMock.js"
+    }
   },
   "keywords": [],
   "author": "",
@@ -16,12 +33,20 @@
   },
   "homepage": "https://github.com/eroiyo/WebPack-BoilerPlate#readme",
   "dependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.14.5",
+    "babel": "^6.23.0",
+    "identity-obj-proxy": "^3.0.0",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.14.5",
+    "@babel/core": "^7.14.8",
+    "@babel/node": "^7.14.7",
+    "hint": "^6.1.4",
     "@fortawesome/fontawesome-free": "^5.15.3",
     "css-loader": "^5.2.7",
     "html-webpack-plugin": "^5.3.2",
+    "jest": "^27.0.6",
     "style-loader": "^3.1.0",
     "webpack": "^5.44.0",
     "webpack-cli": "^4.7.2",
